SMGCPA: Fix offset type
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/SMG-has-value-edges-invariant@31819 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/SMGState.java
@@ -1579,7 +1579,7 @@ public class SMGState implements UnmodifiableSMGState, AbstractQueryableState, G
    * @param smgObject The memory the given Address belongs to.
    * @return returns a possible new State
    */
-  protected SMGState free(Integer offset, SMGObject smgObject) throws SMGInconsistentException {
+  protected SMGState free(long offset, SMGObject smgObject) throws SMGInconsistentException {
 
     if (!heap.isHeapObject(smgObject) && !heap.isObjectExternallyAllocated(smgObject)) {
       // You may not free any objects not on the heap.
